Update chat.tsx
Hi @gotochkin 
I know this is a last min request. I was thinking may be just for a couple of days, could we include this URL in the chat - to enable users to snap a toy placed in the demo booth and copy the description to the chat input. Kindly let me know if this is ok to just merge this PR:

   <a href="https://toys-uxu5wi2jpa-uc.a.run.app" target="_blank">Toys like this: Click and Copy!
            </a>